+++ NEWS.md
@@ -1,5 +1,6 @@
 # coronavirus 0.0.3
 
+- Connect all charts in dxy and weixin tabs
 - Corrected deaths and recovered in dianxiangying data, was swapped, see [#2](https://github.com/JohnCoene/coronavirus/issues/2)
 - Number of suspected by city given by DianXiangYing is wildly inaccurate, has been removed.